Наконец убрал дурацкий move файла после скачки

+++ src/main/java/ru/gadjini/telegram/smart/bot/commons/service/file/temp/TempFileService.java
@@ -142,6 +142,9 @@ public class TempFileService {
     }
 
     public void delete(SmartTempFile file) {
+        if (file == null) {
+            return;
+        }
         try {
             if (isRemoteFile(file.getAbsolutePath())) {
                 contentApi.delete(file);
